+++ jest.config.js
@@ -38,5 +38,5 @@ module.exports = {
 
   // The glob patterns Jest uses to detect test files
   testMatch: ["<rootDir>/test/**/*.spec.js"],
-  // testMatch: ["<rootDir>/test/**/one-level/element-types-template.spec.js"],
+  // testMatch: ["<rootDir>/test/**/one-level/external-template.spec.js"],
 };